requirements.txt
asttokens==2.1.0
azure-common==1.1.28
azure-communication-email==1.0.0b1
azure-core==1.26.1
azure-functions==1.12.0
azure-mgmt-core==1.3.2
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.5
jupyter_core==5.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
msrest==0.7.1
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.6
pywhatkit==5.3
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wikipedia==1.4.0
zipp==3.10.0
